speech/python/requirements.txt
azure-cognitiveservices-speech==1.30.0
python-dotenv==1.0.1
